C09 - ex00 done

+++ C09/ex00/BitcoinExchange.hpp
@@ -13,7 +13,7 @@
 class BitcoinExchange {
 
 	private:
-		std::map <int, int>							database;
+		std::map <int, float>						database;
 		std::multimap <std::string, std::string>	input;
 
 	public:
@@ -27,11 +27,9 @@ class BitcoinExchange {
 		void	print_db();
 		void	print_input();
 		double	change(std::string str1, std::string str2);
-
-	
 };
 
-std::map<int, int>						ft_createdb();
+std::map<int, float>					ft_createdb();
 std::multimap<std::string, std::string>	ft_createinput(char *str);
 bool									ft_checkFirst(std::string *str);
 bool									ft_checkSecond(std::string *str);
